refactor: tooltip
